The INI override folder can now be overridden
By setting the global variable 'EZP_INI_OVERRIDE_FOLDERS' the
override folders can be changed to whatever the site want.
The folder path is relative from the document root, not inside the
eZ publish settings folder.

This means that ini settings belonging to the site can be moved
outside of the eZ publish folders.