Add support for PSR-3 loggers in eZDebug
The PSR-3 loggers will then override the internal logging
mechanism, any log message is instead passed to the logger.

The logger must be created using a logger factory, it
is a callable that creates the log instances when needed.
The factory will receive a log level name as parameter and
must return a log instance, or return null to disable logging
for that level.

The factory is configured by setting the global variable
$GLOBALS["ezpDebug"]["loggerFactory]