Fix warning caused by passing non-object to get_class
As of PHP7.2 the `get_class` function raises a warning when the passed
parameter is a non-object. By calling `is_object` before `get_class` we
ensure that the generated template code can't result in a warning
from `get_class`. And by defaulting to an empty string we avoid
potential future issues with calling `strtolower` on a boolean.
https://www.php.net/manual/en/function.get-class.php